+++ src/cfCore.ts
@@ -47,6 +47,8 @@ export class CfCore {
 
         context.subscriptions.push(commands.registerCommand('f5-cf.inspect', async () => {
 
+            ext.telemetry.capture({ command: 'f5-cf.inspect' });
+
             await window.withProgress({
                 location: ProgressLocation.Notification,
                 title: `Getting CF inspect`
@@ -62,6 +64,8 @@ export class CfCore {
 
         context.subscriptions.push(commands.registerCommand('f5-cf.getDec', async () => {
 
+            ext.telemetry.capture({ command: 'f5-cf.getDec' });
+
             await window.withProgress({
                 location: ProgressLocation.Notification,
                 title: `Getting CF Dec`
@@ -76,6 +80,8 @@ export class CfCore {
         }));
 
         context.subscriptions.push(commands.registerCommand('f5-cf.postDec', async () => {
+
+            ext.telemetry.capture({ command: 'f5-cf.postDec' });
 
             await window.withProgress({
                 location: ProgressLocation.Notification,
@@ -105,6 +111,8 @@ export class CfCore {
 
         context.subscriptions.push(commands.registerCommand('f5-cf.getTrigger', async () => {
 
+            ext.telemetry.capture({ command: 'f5-cf.getTrigger' });
+
             await window.withProgress({
                 location: ProgressLocation.Notification,
                 title: `Getting CF Trigger details`
@@ -121,6 +129,7 @@ export class CfCore {
 
         context.subscriptions.push(commands.registerCommand('f5-cf.triggerDryRun', async () => {
 
+            ext.telemetry.capture({ command: 'f5-cf.triggerDryRun' });
             await window.withProgress({
                 location: ProgressLocation.Notification,
                 title: `Posting CF Trigger as Dry-Run`
@@ -133,6 +142,8 @@ export class CfCore {
         }));
 
         context.subscriptions.push(commands.registerCommand('f5-cf.trigger', async () => {
+
+            ext.telemetry.capture({ command: 'f5-cf.trigger' });
 
             await window.withProgress({
                 location: ProgressLocation.Notification,
@@ -147,6 +158,7 @@ export class CfCore {
 
         context.subscriptions.push(commands.registerCommand('f5-cf.reset', async () => {
 
+            ext.telemetry.capture({ command: 'f5-cf.reset' });
             await window.withProgress({
                 location: ProgressLocation.Notification,
                 title: `Posting CF RESET`