+++ src/devicesCore.ts
@@ -95,7 +95,7 @@ export default function devicesCore(context: ExtensionContext, f5OutputChannel: 
 
     context.subscriptions.push(commands.registerCommand('f5.connectDevice', async (device) => {
 
-        if (connecting) { 
+        if (connecting) {
             return;
         } else {
             connecting = true;
@@ -164,6 +164,18 @@ export default function devicesCore(context: ExtensionContext, f5OutputChannel: 
 
                 logger.info('F5 Connect Discovered', connect);
 
+
+                // ts-todo: add details about platformMarketinName/VE/m
+                ext.telemetry.capture({
+                    command: "f5.connectDevice",
+                    connect: {
+                        product: connect.product,
+                        version: connect.version,
+                        atc: connect.atc,
+                        baseRegKey: ext.f5Client?.host?.license.registrationKey
+                    }
+                });
+
                 ext.hostsTreeProvider.connectedDevice = ext.f5Client;
                 ext.hostsTreeProvider.refresh();
 
@@ -176,22 +188,30 @@ export default function devicesCore(context: ExtensionContext, f5OutputChannel: 
                 logger.error('Connect/Discover failed', err);
             });
 
+        // const x = ext.f5Client;
         wait(1000);
         connecting = false;
     }));
 
     context.subscriptions.push(commands.registerCommand('f5.getProvider', async () => {
         const x = ext.f5Client;
+        ext.telemetry.capture({ command: "f5.getProvider" });
+
         ext.f5Client?.https('/mgmt/tm/auth/source')
             .then(resp => ext.panel.render(resp));
     }));
 
 
     context.subscriptions.push(commands.registerCommand('f5.getF5HostInfo', async () => {
+
+        ext.telemetry.capture({ command: "f5.getF5HostInfo" });
+
         ext.panel.render(ext.f5Client?.host);
     }));
 
     context.subscriptions.push(commands.registerCommand('f5.disconnect', () => {
+
+        ext.telemetry.capture({ command: "f5.disconnect" });
 
         if (ext.f5Client) {
             ext.f5Client.disconnect();
@@ -204,7 +224,10 @@ export default function devicesCore(context: ExtensionContext, f5OutputChannel: 
     }));
 
     context.subscriptions.push(commands.registerCommand('f5.clearPassword', async (item) => {
-        return ext.hostsTreeProvider.clearPassword(item.label);
+        ext.telemetry.capture({ command: "f5.clearPassword" });
+        const device = item?.label ? item.label : item?.device ? item.device : item;
+        logger.error(`Clearing password for ${device}`);
+        return ext.hostsTreeProvider.clearPassword(device);
     }));
 
 
@@ -213,6 +236,7 @@ export default function devicesCore(context: ExtensionContext, f5OutputChannel: 
     }));
 
     context.subscriptions.push(commands.registerCommand('f5.codeJsonCfg', async () => {
+        ext.telemetry.capture({ command: "f5.codeJsonCfg" });
         commands.executeCommand('workbench.action.openSettingsJson');
     }));
 
@@ -225,6 +249,9 @@ export default function devicesCore(context: ExtensionContext, f5OutputChannel: 
     }));
 
     context.subscriptions.push(commands.registerCommand('f5.editHost', async (hostID) => {
+
+        ext.telemetry.capture({ command: "f5.editHost" });
+
         return await ext.hostsTreeProvider.editDevice(hostID);
     }));
 
@@ -283,6 +310,9 @@ export default function devicesCore(context: ExtensionContext, f5OutputChannel: 
 
     context.subscriptions.push(commands.registerCommand('f5.deviceImport', async (item) => {
 
+        // ts-todo: make sure this is only called by user
+        ext.telemetry.capture({ command: "f5.deviceImport" });
+
         return utils.getText()
             .then(async text => {
                 await deviceImport(text);
@@ -296,6 +326,9 @@ export default function devicesCore(context: ExtensionContext, f5OutputChannel: 
 
 
     context.subscriptions.push(commands.registerCommand('f5.createUCS', async () => {
+
+        ext.telemetry.capture({ command: 'f5.createUCS' });
+
         // create ucs on f5
 
         // https://code.visualstudio.com/api/references/icons-in-labels
@@ -327,6 +360,8 @@ export default function devicesCore(context: ExtensionContext, f5OutputChannel: 
 
     context.subscriptions.push(commands.registerCommand('f5.deleteUCS', async (item) => {
 
+        ext.telemetry.capture({ command: 'f5.deleteUCS' });
+
         return await window.withProgress({
             location: { viewId: 'ipView' },
         }, async () => {
@@ -345,6 +380,8 @@ export default function devicesCore(context: ExtensionContext, f5OutputChannel: 
 
     context.subscriptions.push(commands.registerCommand('f5.downloadUCS', async (filename) => {
         // download ucs from f5
+
+        ext.telemetry.capture({ command: 'f5.downloadUCS' });
 
         return await window.withProgress({
             location: ProgressLocation.Window,
@@ -388,6 +425,8 @@ export default function devicesCore(context: ExtensionContext, f5OutputChannel: 
      */
 
     context.subscriptions.push(commands.registerCommand('f5.installRPM', async (selectedRPM) => {
+
+        ext.telemetry.capture({ command: 'f5.installRPM' });
 
         const downloadResponses = [];
         const upLoadResponses = [];
@@ -484,6 +523,8 @@ export default function devicesCore(context: ExtensionContext, f5OutputChannel: 
     }));
 
     context.subscriptions.push(commands.registerCommand('f5.unInstallRPM', async (rpm) => {
+
+        ext.telemetry.capture({ command: 'f5.unInstallRPM' });
 
         let prompt: boolean = true;