+++ src/extensionVariables.ts
@@ -35,6 +35,8 @@ import { F5Client } from "./f5Client";
 import { AtcVersions, AtcVersionsClient, ExtHttp } from 'f5-conx-core';
 import { AS3TreeProvider } from './treeViewsProviders/as3TreeProvider';
 import { F5TreeProvider } from './treeViewsProviders/hostsTreeProvider';
+import { Telemetry } from './telemetry';
+import { XcDiag } from './tmosXcDiag';
 
 type KeyTar = typeof keyTarType;
 
@@ -48,7 +50,9 @@ export namespace ext {
     export let extHttp: ExtHttp;
     export let keyTar: KeyTar;
     export let hostsTreeProvider: F5TreeProvider;
+    export let telemetry: Telemetry;
     export let as3Tree: AS3TreeProvider;
+    export let xcDiag: XcDiag;
     export let eventEmitterGlobal: EventEmitter;
     export let atcVersions: AtcVersions;
     export let connectBar: StatusBarItem;
@@ -88,8 +92,12 @@ export async function initSettings(context: ExtensionContext) {
     // todo: setup settings for external http proxy - should probably set environment vars
     ext.eventEmitterGlobal = new EventEmitter();
 
+    // ext.tele Telemetry
+
+    ext.teemAgent = `${context.extension.packageJSON.name}/${context.extension.packageJSON.version}`;
+
     ext.cacheDir = path.join(ext.context.extensionPath, 'cache');
-    process.env.F5_CONX_CORE_EXT_HTTP_AGENT = 'The F5 VScode Extension';
+    process.env.F5_CONX_CORE_EXT_HTTP_AGENT = ext.teemAgent;
     process.env.F5_CONX_CORE_CACHE = ext.cacheDir;
 
     // process.env.NODE_TLS_REJECT_UNAUTHORIZED = "0";
@@ -104,10 +112,10 @@ export async function initSettings(context: ExtensionContext) {
         .on('log-info', msg => logger.info(msg))
         .on('log-warn', msg => logger.warn(msg))
         .on('log-error', msg => logger.error(msg))
-        .on('failedAuth', msg => {
+        .on('failedAuth', async msg => {
             window.showErrorMessage('Failed Authentication - Please check password');
             logger.error('Failed Authentication Event!', ext.f5Client?.device, msg);
-            ext.f5Client?.clearPassword();
+            await ext.f5Client?.clearPassword(ext.f5Client?.device.device);
             commands.executeCommand('f5.disconnect');
         });